feat: comprehensive update — AI services, auth, payments, frontend pages, deployment config
Backend:
- Enhanced AI services: advanced matching, fraud detection, skill analyzer, services API
- Improved auth: social login, token refresh, multi-factor support
- Payment domain: multi-currency, Stripe improvements, wallet, disputes
- Chat: messages API, WebSocket enhancements
- Models: dispute, invoice, milestone, refund, scope_change, time_entry
- Upload and availability calendar improvements
- Dockerfile optimizations

Frontend:
- New/updated landing pages: compare, cost-calculator, faq, features, tools, why-hire, post-project
- Messaging: ChatInbox and RealtimeChat improvements
- Global CSS and layout refinements
- Removed old page.tsx, added frontend Dockerfile
- Next.js config and Vercel config updates

Infrastructure:
- Updated DO app-production spec with proper env vars and build config
- CI/CD workflow improvements
- Docker compose updates
- Terraform workflow fixes

+++ backend/app/models/dispute.py
@@ -4,6 +4,7 @@ from sqlalchemy import String, Text, DateTime, ForeignKey, Numeric
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 from app.db.base import Base
 from datetime import datetime, timezone
+from decimal import Decimal
 from typing import Optional, TYPE_CHECKING
 import logging
 import enum
@@ -48,7 +49,7 @@ class Dispute(Base):
     created_at: Mapped[datetime] = mapped_column(DateTime, default=lambda: datetime.now(timezone.utc), index=True)
     resolved_at: Mapped[datetime] = mapped_column(DateTime, nullable=True)
     resolution: Mapped[str] = mapped_column(Text, nullable=True)
-    resolution_amount: Mapped[float] = mapped_column(Numeric(12, 2), nullable=True)
+    resolution_amount: Mapped[Decimal] = mapped_column(Numeric(12, 2), nullable=True)
     updated_at: Mapped[datetime] = mapped_column(DateTime, default=lambda: datetime.now(timezone.utc), onupdate=lambda: datetime.now(timezone.utc))
 
     # Relationships
