+++ backend/app/models/milestone.py
@@ -4,6 +4,7 @@ from sqlalchemy import String, Integer, Text, Numeric, DateTime, ForeignKey
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 from app.db.base import Base
 from datetime import datetime, timezone
+from decimal import Decimal
 from typing import TYPE_CHECKING
 import logging
 import enum
@@ -31,7 +32,7 @@ class Milestone(Base):
     contract_id: Mapped[int] = mapped_column(ForeignKey("contracts.id"), index=True)
     title: Mapped[str] = mapped_column(String(255))
     description: Mapped[str] = mapped_column(Text)
-    amount: Mapped[float] = mapped_column(Numeric(12, 2))
+    amount: Mapped[Decimal] = mapped_column(Numeric(12, 2))
     due_date: Mapped[datetime] = mapped_column(DateTime)
     status: Mapped[str] = mapped_column(String(20), default=MilestoneStatus.PENDING.value, index=True)
     deliverables: Mapped[str] = mapped_column(Text, nullable=True)  # JSON string
