+++ backend/app/models/scope_change.py
@@ -3,6 +3,7 @@ from sqlalchemy import String, Numeric, DateTime, Text, ForeignKey
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 from app.db.base import Base
 from datetime import datetime, timezone
+from decimal import Decimal
 from typing import TYPE_CHECKING
 
 if TYPE_CHECKING:
@@ -19,8 +20,8 @@ class ScopeChangeRequest(Base):
     description: Mapped[str] = mapped_column(Text)
     reason: Mapped[str] = mapped_column(Text, nullable=True)
     status: Mapped[str] = mapped_column(String(20), default="pending") # pending, approved, rejected, cancelled
-    old_amount: Mapped[float] = mapped_column(Numeric(12, 2), nullable=True)
-    new_amount: Mapped[float] = mapped_column(Numeric(12, 2), nullable=True)
+    old_amount: Mapped[Decimal] = mapped_column(Numeric(12, 2), nullable=True)
+    new_amount: Mapped[Decimal] = mapped_column(Numeric(12, 2), nullable=True)
     old_deadline: Mapped[datetime] = mapped_column(DateTime, nullable=True)
     new_deadline: Mapped[datetime] = mapped_column(DateTime, nullable=True)
     created_at: Mapped[datetime] = mapped_column(DateTime, default=lambda: datetime.now(timezone.utc))
